Clarify instructions
When I did this problem, I enclosed an evenIterable and a distinct oddIterable in the closure of generate(). Then I swapped between them when next was called with swap. This results in the following behavior:

$ var iterable = generate(true);
$ iterable.next(); // 2
$ iterable.next(); // 4
$ iterable.next(true); // 1; the program expects 5
$ iterable.next(); // 3
$ iterable.next(true); // 6; the program would expect 4 given the previous line

I am unclear about whether I misunderstood or the terminology is incorrect. These few words for an example may help other people who got confused like I did.

Thanks for the course!